Added scroll support
CHANGED:
* Using `size_t` for cursor positioning

ADDED:
* `cursor_move` moves cursor as close to provided `n` as possible

+++ src/cursor.h
@@ -1,14 +1,16 @@
 #pragma once
 
+#include <stddef.h>
+
 enum cursor_direction { CURSOR_UP, CURSOR_DOWN, CURSOR_RIGHT, CURSOR_LEFT };
 
 struct cursor {
-	int x;
-	int x_min;
-	int x_max;
-	int y;
-	int y_min;
-	int y_max;
+	size_t x;
+	size_t x_min;
+	size_t x_max;
+	size_t y;
+	size_t y_min;
+	size_t y_max;
 };
 
-void cursor_move(struct cursor *, enum cursor_direction, int);
+void cursor_move(struct cursor *, enum cursor_direction, size_t);
